move Juno into an extension
